third_party/WebKit/LayoutTests/external/wpt/WebIDL/ecmascript-binding/iterator-prototype-object-expected.txt

This is a testharness.js-based test.
PASS Has %IteratorPrototype% as prototype
PASS next() exists and is writable, enumerable, and configurable
FAIL next() throws TypeError when called on ineligible receiver assert_throws: function "() => {
    iteratorProto.next.call(new Headers().entries());
  }" did not throw
FAIL Object.prototype.toString returns correct value assert_equals: expected "[object URLSearchParams Iterator]" but got "[object Iterator]"
FAIL @@toStringTag has correct value assert_equals: expected "URLSearchParams Iterator" but got "Iterator"
FAIL Is specific to an interface assert_not_equals: got disallowed value object "[object Iterator]"
Harness: the test ran to completion.
